+++ M1/Lat2/tabung.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #define phi 3.14
 
-int r, v, t;
+float r, v, t;
 
 //Fungsi formula volume tabung
-int vTabung(int r, int t){
+float vTabung(float r, float t){
     v = phi * r * r * t;
 
     return v;
@@ -13,9 +13,9 @@ int vTabung(int r, int t){
 void atabung(){
     printf("Inputkan nilai jari-jari dan tinggi [dalam cm] \n");
     printf("Jari-jari : ");
-    scanf("%d", &r);
+    scanf("%f", &r);
     printf("Tinggi : ");
-    scanf("%d", &t);    
+    scanf("%f", &t);    
     vTabung(r, t);
-    printf("Volume : %d  lingkaran dengan Jari-jari : %d dan tinggi : %d\n\n", v, r, t);
+    printf("Volume : %.2f  lingkaran dengan Jari-jari : %.2f dan tinggi : %.2f\n\n", v, r, t);
 }
